feat(melody): Scale melody from melody popup.
